update cli for running georeference/thumb ops
